# Improve performance (real and perceived) of Console terminal input