--- a/packages/bvaughn-architecture-demo/components/sources/AutoComplete/AutoCompleteList.tsx
+++ b/packages/bvaughn-architecture-demo/components/sources/AutoComplete/AutoCompleteList.tsx
@@ -174,6 +174,10 @@ export default function AutoCompleteListOuter({
     selectedIndex,
   };
 
+  if (matches === null) {
+    return null;
+  }
+
   return (
     <Popup
       clientX={cursorClientX}