--- a/packages/bvaughn-architecture-demo/components/sources/AutoComplete/utils/findMatches.test.ts
+++ b/packages/bvaughn-architecture-demo/components/sources/AutoComplete/utils/findMatches.test.ts
@@ -77,4 +77,15 @@ describe("findMatches", () => {
       findMatches("foo", "bar", null, createPropertiesFromNames("bbaaarr", "abarr", "cbar"))
     ).toEqual(["cbar", "abarr", "bbaaarr"]);
   });
+
+  it("should escape special characters", () => {
+    expect(
+      findMatches(
+        null,
+        "fb(",
+        null,
+        createPropertiesFromNames("fooBar()", "fooBarBaz()", "barBaz()")
+      )
+    ).toEqual(["fooBar()", "fooBarBaz()"]);
+  });
 });