… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/console/should-be-searchable-on-complex-content.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/console/should-be-searchable.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/console/should-evaluate-terminal-expressions-at-an-execution-point.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/console/should-show-a-button-to-clear-terminal-expressions.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/console/should-show-and-hide-search-input-when-Enter-and-Escape-are-typed.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/console/should-show-the-context-menu-on-top-of-other-messages-and-the-current-time-indicator.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/console/should-support-seeking-to-a-message-execution-point.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/source-and-console/should-auto-suggestion-text-based-on-the-current-expression.json
… packages/bvaughn-architecture-demo/playwright/protocol-fixtures/source-and-console/should-update-the-current-time-when-the-next-previous-log-point-buttons-are-clicked.json
… packages/bvaughn-architecture-demo/playwright/snapshots/tests/source-and-console.ts-snapshots/dark/log-point-auto-complete-list-window-linux.png
… packages/bvaughn-architecture-demo/playwright/snapshots/tests/source-and-console.ts-snapshots/light/log-point-auto-complete-list-window-linux.png