This commit has three significant change:

| Change | Improves |
| :--- | :--- |
| Moved the `prevMatchingTextIndex` increment inside of the "did match" conditional | ranking relevance |
| Replaced `localeCompare` to a `RegExp` for fuzzy match comparison | performance |
| Pre-fetch properties on _focus_ rather than waiting for the first character to be typed | perceived performance |

I explain each of these changes below in greater detail, but here's a quick before-and-after demo video:

| Before | After |
| ------------- | ------------- |
| ![Kapture 2022-11-21 at 16 42 40](https://user-images.githubusercontent.com/29597/203163951-db023e57-5da5-4d5f-94a3-d057c7a148c8.gif) | ![Kapture 2022-11-21 at 16 38 56](https://user-images.githubusercontent.com/29597/203163747-755524b9-10db-4ee1-83cb-4697d9a8cc42.gif) |

# 1. Moving prev-matching-index
This was an accident  when I initially wrote the weight calculation. It was basically ranking all matches equally in this regard. The images below show an example of how the location change improved result relevancy.

| Before | After |
| ------------- | ------------- |
| <img width="241" alt="Screen Shot 2022-11-21 at 4 32 10 PM" src="https://user-images.githubusercontent.com/29597/203164463-bb333a82-5a5f-4676-b327-2427dd25253f.png">  | <img width="243" alt="Screen Shot 2022-11-21 at 4 32 06 PM" src="https://user-images.githubusercontent.com/29597/203164459-8c89fc3a-30be-4ad7-9a63-a09196a48cd2.png"> |

# 2. Replacing `localeCompare` with a `RegExp`
I should have benchmarked this more initially. Using `localeCompare` was silly of me. It's crazy slow. Surprisingly though, `RegExp` turned out to be a lot faster than even `toLowerCase` and `===`. I benchmarked all three and found the following numbers– although the specific results will vary a bit based on the shape of the data (how many properties, what the search term is):

* `localeCompare` averages **588** ops/sec.
* `toLowerCase` followed by an equality check averages **146,616** ops/sec
* Converting the needle to a `RegExp` and checking the match averages **430,936** ops/sec.

# 3. Pre-fetching preview data
This change is a bit subjective. It improves the _perceived_ performance at the cost of (potentially) fetching a little extra data. Before we were waiting until the user typed at least one character before we fetched preview data to show in the typeahead. Now we pre-fetch on focus, so when the user types we can begin searching sooner.

This performance trace illustrates the amount of time we spent _waiting for data_ before:
<img width="531" alt="Screen Shot 2022-11-21 at 4 33 34 PM" src="https://user-images.githubusercontent.com/29597/203166345-d87e8871-a6c7-4d50-806d-52fe67662a8f.png">